+++ CHANGES.rst
@@ -1,11 +1,23 @@
 Unreleased
 ----------
 
+- Fix Schedule block
+  [jchandelle]
+
+- Fix start and end dates to include timezone in ical view
+  [mpeeters]
+
+- Adapt session serializer to handle timezone correctly
+  [mpeeters]
+
+- Fix Schedule Block default value and trad
+  [jchandelle]
+
+- Add Schedule Block
+  [jchandelle]
+
 - Fix relationList empty error if editing an object with "presenters" field.
   [sneridagh]
-
-2022-05-21
-----------
 
 - Add spaces between block
   [jchandelle]
@@ -27,6 +39,9 @@ Unreleased
 
 - implement design and related blocks
   [jchandelle, tlambert, duchenean]
+
+2022-05-21
+----------
 
 - add notes and userful docker commands
   [tkimnguyen]